Update backup-nessus.service
The Nessus backup service creates a local backup named nessus.tgz, then deletes the container created for the backup after it is run, as opposed to leaving it as I was doing previously. This allows for changes to the backup service to be implemented when the config file (backup-nessus.service) is altered.